requirements.txt
opencv-python~=4.11.0.86
numpy<2.0.0
pygetwindow~=0.0.9
pyautogui~=0.9.54
ultralytics==8.3.111
contourpy==1.2.0
mss~=10.0.0
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==2.6.0+cu118
pywin32
win32gui
pytesseract~=0.3.13
pyserial~=3.5